Refactored for improved thread safety
Dynamically allocated BleSerial objects are now stored in a fixed-length array of SharedPtr objects. The size of this array can be set by the application configuration and defaults to the maximum number of connections configured for the stack.

SharedPtr is now being used to pass references to the BleSerial objects. The lifetime of the SharedPtr is no longer controlled by the UARTService object which makes it possible for other threads to check if a particular BleSerial object has been shutdown. Previously, all shared references to the BleSerial object would be invalidated (and set to nullptr) and could cause a deadlock and/or crashing in such a scenario if another thread was still reading/writing from the BleSerial.