Introduce InvertedDigitalOut subclass
This commit introduces a subclass of Mbed's `DigitalOut` class, an `InvertedDigitalOut`. This enables the following use case:

An API, such as a device driver with a power shutdown pin for example, may take in a pointer to an `mbed::interface::DigitalOut` instance to control the external power shutdown signal. The API has no way of knowing if the power shutdown signal is active high or active low. The typical design pattern to overcome this is to include an additional argument, eg: `bool shutdown_is_active_low`, complicating the API. This is a common enough scenario that I think it is worth creating a more straemlined workaround for it.

With this class the user may simply provide a pointer to an `InvertedDigitalOut` instance instead. The driver then does not need to know the polarity of the control signal. It is a fairly lightweight and straightforward solution.